requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
humanize==2.4.0
idna==2.9
kaitaistruct==0.9
libarchive==0.4.7
nose==1.3.7
ObjDict==0.4.4
Pillow==7.1.2
pkg-resources==0.0.0
pycdlib==1.9.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tqdm==4.46.0
urllib3==1.25.9
